+++ MovieNight/Controllers/MoviesController.cs
@@ -124,16 +124,31 @@ namespace MovieNight.Controllers
         // GET: Movies
         public ActionResult Manage()
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             return View(db.Movies.ToList());
         }
 
         public ActionResult SearchApi(string query)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             return this.Json(client.SearchMovie(query).Results);
         }
 
         public ActionResult GetDetailsFromApi(int id)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             TMDbLib.Objects.Movies.Movie result = client.GetMovie(id);
             Crew Director = client.GetMovieCredits(id).Crew.Where(crew => crew.Job == "Director").FirstOrDefault();
 
@@ -193,44 +208,15 @@ namespace MovieNight.Controllers
         // GET: Movies/Create
         public ActionResult Create()
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             ViewBag.DirectorID = new SelectList(db.Directors, "ID", "Name");
             return View();
         }
-
-        public async System.Threading.Tasks.Task<ActionResult> Search(string title)
-        {
-            string _address = "http://www.omdbapi.com/?s=" + title + "&type=movie&r=xml";
-            try
-            {
-                ViewBag.Movies = new List<string>();
-                HttpWebRequest request = WebRequest.Create(_address) as HttpWebRequest;
-                HttpWebResponse response = request.GetResponse() as HttpWebResponse;
-
-                XmlDocument xmlDoc = new XmlDocument();
-                xmlDoc.Load(response.GetResponseStream());
-
-                XmlNamespaceManager nsmgr = new XmlNamespaceManager(xmlDoc.NameTable);
-                nsmgr.AddNamespace("rest", "http://schemas.microsoft.com/search/local/ws/rest/v1");
-
-                //Get all locations in the response and then extract the formatted address for each location
-                XmlNodeList moviesElement = xmlDoc.SelectNodes("//rest:result", nsmgr);
-                foreach (XmlNode movie in moviesElement)
-                {
-                    ViewBag.Movies.Add(movie.SelectSingleNode(".//rest:title", nsmgr).InnerText);
-                }
-
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e.Message);
-
-                Console.Read();
-                return null;
-            }
-            
-            return View();
-        }
-
+        
         // POST: Movies/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
@@ -238,11 +224,16 @@ namespace MovieNight.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,Title,ReleaseDate,Genre,Plot,DirectorID,Rating,Poster,Trailer")] MovieNight.Models.Movie movie)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Movies.Add(movie);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Manage");
             }
 
             ViewBag.DirectorID = new SelectList(db.Directors, "ID", "Name", movie.DirectorID);
@@ -252,6 +243,11 @@ namespace MovieNight.Controllers
         // GET: Movies/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -272,11 +268,16 @@ namespace MovieNight.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID,Title,ReleaseDate,Genre,Plot,DirectorID,Rating,Poster,Trailer")] MovieNight.Models.Movie movie)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(movie).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Manage");
             }
             ViewBag.DirectorID = new SelectList(db.Directors, "ID", "Name", movie.DirectorID);
             return View(movie);
@@ -285,6 +286,11 @@ namespace MovieNight.Controllers
         // GET: Movies/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -302,10 +308,15 @@ namespace MovieNight.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["user"] == null || (Session["user"] as User).Role != Role.Admin)
+            {
+                return RedirectToAction("AccessDenied", "Users");
+            }
+
             MovieNight.Models.Movie movie = db.Movies.Find(id);
             db.Movies.Remove(movie);
             db.SaveChanges();
-            return RedirectToAction("Index");
+            return RedirectToAction("Manage");
         }
 
         protected override void Dispose(bool disposing)
